feat: update the styling for the error threshold
